refactor: ホームスクリーンのロジックをカスタムフックに移行
- 認証状態管理を `useAuth` フックに統合
- 習慣管理ロジックを `useHabits` フックに抽出
- コンポーネントの責務を分離し、ロジックを再利用可能に
- 状態管理と副作用を簡潔なカスタムフックに集約
