+++ app/(tabs)/index.tsx
@@ -1,356 +1,37 @@
 import Auth from '@/components/Auth';
-import { supabase } from '@/lib/supabase';
-import { useEffect, useState, useCallback } from 'react';
-import { Session } from '@supabase/supabase-js';
+import { useAuth } from '@/hooks/useAuth';
+import { useHabits } from '@/hooks/useHabits';
 import { VStack } from '@/components/ui/vstack';
 import AchievementModal from '../../components/AchievementModal';
 import { Box } from '@/components/ui/box';
 import { HabitItem } from '@/components/HabitItem';
 import { WeekView } from '@/components/WeekView';
-import { Audio } from 'expo-av';
 import NewHabitModal from '@/components/NewHabitModal';
 import { Text } from '@/components/ui/text';
 import HabitFab from '@/components/HabitFab';
 import { ScrollView } from 'react-native';
-import * as Haptics from 'expo-haptics';
-import { updatePushToken } from '@/utils/notifications';
 
 export default function Index() {
-  const [session, setSession] = useState<Session | null>(null);
+  // 認証状態を取得
+  const { session } = useAuth();
 
-  useEffect(() => {
-    supabase.auth.getSession().then(({ data: { session } }) => {
-      setSession(session);
-      if (session?.user) {
-        updatePushToken();
-      }
-    });
-
-    supabase.auth.onAuthStateChange((_event, session) => {
-      setSession(session);
-      if (session?.user) {
-        updatePushToken();
-      }
-    });
-  }, []);
-
-  function getMaxConsecutiveDays(dates: string[]): number {
-    const sorted = [...dates].sort();
-    let maxStreak = 0;
-    let currentStreak = 0;
-    let prevDate: Date | null = null;
-
-    for (const dateStr of sorted) {
-      const dateObj = new Date(dateStr);
-      if (
-        prevDate &&
-        dateObj.getTime() - prevDate.getTime() === 24 * 60 * 60 * 1000
-      ) {
-        currentStreak++;
-      } else {
-        currentStreak = 1;
-      }
-      maxStreak = Math.max(maxStreak, currentStreak);
-      prevDate = dateObj;
-    }
-
-    return maxStreak;
-  }
-
-  // 追加: 現在の日付を含めた連続日数を計算する関数
-  function getCurrentConsecutiveDays(dates: string[]): number {
-    const dateSet = new Set(dates);
-    let currentStreak = 0;
-    let date = new Date();
-    // 日付フォーマットは "YYYY-MM-DD" で比較
-    while (dateSet.has(date.toISOString().split('T')[0])) {
-      currentStreak++;
-      date.setDate(date.getDate() - 1);
-    }
-    return currentStreak;
-  }
-
-  type Habit = {
-    id: string;
-    name: string;
-    streak: number;
-    completedDates: string[];
-    totalDays: number;
-    is_public: boolean;
-    achieved_at?: string;
-    goal?: number;
-  };
-
-  const [habits, setHabits] = useState<Habit[]>([]);
-  const [newHabitModalData, setNewHabitModalData] = useState<{
-    isOpen: boolean;
-    habitName: string;
-  }>({
-    isOpen: false,
-    habitName: '',
-  });
-
-  const [achievementData, setAchievementData] = useState<{
-    isOpen: boolean;
-    streak: number;
-    habitName: string;
-  }>({
-    isOpen: false,
-    streak: 0,
-    habitName: '',
-  });
-
-  // habitデータの取得
-  const fetchHabits = useCallback(async () => {
-    try {
-      const { data, error } = await supabase
-        .from('habits')
-        .select('*')
-        .eq('user_id', session?.user?.id)
-        .order('created_at', { ascending: true });
-
-      if (error) throw error;
-
-      if (data) {
-        const formattedData = data.map((habit) => ({
-          id: habit.id,
-          name: habit.name,
-          streak: habit.streak || 0,
-          completedDates: habit.completed_dates || [],
-          totalDays: habit.total_days || 0,
-          is_public: habit.is_public || true,
-          achieved_at: habit.achieved_at,
-          goal: habit.goal,
-        }));
-        setHabits(formattedData);
-      }
-    } catch (error) {
-      console.error('Error fetching habits:', error);
-    }
-  }, [session]);
-
-  useEffect(() => {
-    if (session?.user) {
-      fetchHabits();
-    }
-  }, [session, fetchHabits]);
-
-  // 習慣の追加
-  const handleAddHabit = async (habitName: string) => {
-    if (!habitName.trim() || habitName.length > 16) {
-      return true; // エラーあり
-    }
-    if (habits.some((habit) => habit.name === habitName.trim())) {
-      return true; // エラーあり
-    }
-
-    try {
-      const { data, error } = await supabase
-        .from('habits')
-        .insert([
-          {
-            name: habitName,
-            streak: 0,
-            completed_dates: [],
-            user_id: session?.user?.id,
-          },
-        ])
-        .select()
-        .single();
-
-      if (error) throw error;
-
-      if (data) {
-        setHabits([...habits, data]);
-        setNewHabitModalData({
-          isOpen: true,
-          habitName: habitName,
-        });
-        return false; // エラーなし
-      }
-    } catch (error: any) {
-      console.error('Error adding habit:', error);
-    }
-    return true; // エラーあり
-  };
-
-  // 習慣の完了状態の切り替え
-  const toggleComplete = async (habitId: string, date: string) => {
-    try {
-      // Trigger light feedback for toggling.
-      await Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
-
-      const habit = habits.find((h) => h.id === habitId);
-      if (!habit) {
-        console.error('Habit not found');
-        return;
-      }
-
-      // 日付が有効かチェック
-      const dateObj = new Date(date);
-      if (isNaN(dateObj.getTime())) {
-        console.error('Invalid date:', date);
-        return;
-      }
-
-      const completedDates = habit.completedDates || [];
-      const isCompleted = completedDates.includes(date);
-      const updatedCompletedDates = isCompleted
-        ? completedDates.filter((d) => d !== date)
-        : [...completedDates, date];
-
-      const streak = getMaxConsecutiveDays(updatedCompletedDates);
-
-      // 当日の日付を取得
-      const today = new Date().toISOString().split('T')[0];
-
-      // achieved_at の更新ロジック
-      // 1. 未達成→達成の時だけ更新を検討
-      // 2. 当日の日付の場合のみ
-      // 3. achieved_at が未設定の場合、または達成日が今日でない場合のみ更新
-      let newAchievedAt = habit.achieved_at;
-
-      if (!isCompleted && date === today) {
-        // すでに achieved_at が設定されているか確認
-        const achievedDate = habit.achieved_at
-          ? new Date(habit.achieved_at).toISOString().split('T')[0]
-          : null;
-
-        // achieved_at が未設定、または達成日が今日でない場合のみ更新
-        if (!achievedDate || achievedDate !== today) {
-          newAchievedAt = new Date().toISOString();
-        }
-      }
-
-      // Update habit completion
-      const { data: updatedHabit, error: habitError } = await supabase
-        .from('habits')
-        .update({
-          completed_dates: updatedCompletedDates,
-          streak,
-          total_days: updatedCompletedDates.length,
-          achieved_at: newAchievedAt,
-        })
-        .eq('id', habitId)
-        .eq('user_id', session?.user?.id)
-        .select()
-        .single();
-
-      if (habitError) {
-        console.error('Error updating habit:', habitError);
-        throw habitError;
-      }
-
-      if (!updatedHabit) {
-        console.error('No habit was updated');
-        return;
-      }
-
-      // UI更新とサウンド再生
-      const playSound = async () => {
-        try {
-          const { sound } = await Audio.Sound.createAsync(
-            require('../../assets/sounds/click.mp3'),
-            {
-              shouldPlay: true,
-            },
-          );
-          await sound.playAsync();
-          // サウンドのクリーンアップ
-          return () => {
-            sound.unloadAsync();
-          };
-        } catch (error) {
-          console.error('Error playing sound:', error);
-        }
-      };
-
-      setHabits(
-        habits.map((h) => {
-          if (h.id === habitId) {
-            const today = new Date().toISOString().split('T')[0];
-            if (
-              date === today &&
-              !isCompleted &&
-              updatedCompletedDates.length > 0
-            ) {
-              playSound();
-              const currentStreak = getCurrentConsecutiveDays(
-                updatedCompletedDates,
-              );
-              setAchievementData({
-                isOpen: true,
-                streak: currentStreak, // 現在の日付を含めた連続日数を表示
-                habitName: h.name,
-              });
-            }
-            return {
-              ...h,
-              // DBには過去最高の連続日数を保存（※必要に応じてフィールド名を変更してください）
-              streak: getMaxConsecutiveDays(updatedCompletedDates),
-              completedDates: updatedCompletedDates,
-              totalDays: updatedCompletedDates.length,
-            };
-          }
-          return h;
-        }),
-      );
-    } catch (error: any) {
-      console.error('Error toggling habit:', error.message || error);
-      // On error, trigger error haptic feedback.
-      await Haptics.notificationAsync(Haptics.NotificationFeedbackType.Error);
-      alert('習慣の更新中にエラーが発生しました。もう一度お試しください。');
-    }
-  };
-
-  const editHabitName = async (habitId: string, newName: string) => {
-    try {
-      const { error } = await supabase
-        .from('habits')
-        .update({ name: newName })
-        .eq('id', habitId);
-
-      if (error) throw error;
-
-      // On successful editing, trigger light feedback.
-      await Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
-
-      setHabits(
-        habits.map((habit) =>
-          habit.id === habitId ? { ...habit, name: newName } : habit,
-        ),
-      );
-    } catch (error: any) {
-      console.error('Error updating habit name:', error);
-      // On error, trigger error haptic feedback.
-      await Haptics.notificationAsync(Haptics.NotificationFeedbackType.Error);
-    }
-  };
-
-  const handleDeleteHabit = async (habitId: string) => {
-    try {
-      const { error } = await supabase
-        .from('habits')
-        .delete()
-        .eq('id', habitId);
-
-      if (error) throw error;
-
-      setHabits((prevHabits) =>
-        prevHabits.filter((habit) => habit.id !== habitId),
-      );
-    } catch (error: any) {
-      console.error('Error deleting habit:', error);
-    }
-  };
+  // 習慣データと関連機能を取得
+  const {
+    habits,
+    achievementData,
+    newHabitModalData,
+    fetchHabits,
+    addHabit,
+    toggleComplete,
+    editHabitName,
+    deleteHabit,
+    closeAchievementModal,
+    closeNewHabitModal,
+  } = useHabits(session);
 
   // HabitFabに渡すためのハンドラー
   const handleAddHabitWrapper = async (habitName: string) => {
-    const hasError = await handleAddHabit(habitName);
-    if (!hasError) {
-      // Removed setNewHabit as newHabit is unused
-    }
-    return hasError;
+    return await addHabit(habitName);
   };
 
   return (
@@ -380,7 +61,7 @@ export default function Index() {
                       allHabits={habits}
                       onToggle={(date) => toggleComplete(habit.id, date)}
                       onEdit={(newName) => editHabitName(habit.id, newName)}
-                      onDelete={() => handleDeleteHabit(habit.id)}
+                      onDelete={() => deleteHabit(habit.id)}
                     />
                   ))
                 )}
@@ -388,18 +69,14 @@ export default function Index() {
 
               <AchievementModal
                 isOpen={achievementData.isOpen}
-                onClose={() =>
-                  setAchievementData((prev) => ({ ...prev, isOpen: false }))
-                }
+                onClose={closeAchievementModal}
                 streak={achievementData.streak}
                 habitName={achievementData.habitName}
               />
 
               <NewHabitModal
                 isOpen={newHabitModalData.isOpen}
-                onClose={() =>
-                  setNewHabitModalData((prev) => ({ ...prev, isOpen: false }))
-                }
+                onClose={closeNewHabitModal}
                 habitName={newHabitModalData.habitName}
                 onGoalSet={fetchHabits}
               />
