+++ hooks/useAuth.ts
@@ -1,23 +1,35 @@
-import { useEffect, useState } from 'react';
-import { User } from '@supabase/supabase-js';
+import { useState, useEffect } from 'react';
 import { supabase } from '@/lib/supabase';
+import { Session } from '@supabase/supabase-js';
+import { updatePushToken } from '@/utils/notifications';
 
 export function useAuth() {
-  const [user, setUser] = useState<User | null>(null);
+  const [session, setSession] = useState<Session | null>(null);
 
   useEffect(() => {
+    // 初期セッションの取得
     supabase.auth.getSession().then(({ data: { session } }) => {
-      setUser(session?.user ?? null);
+      setSession(session);
+      if (session?.user) {
+        updatePushToken();
+      }
     });
 
+    // 認証状態変更のリスナー
     const {
       data: { subscription },
     } = supabase.auth.onAuthStateChange((_event, session) => {
-      setUser(session?.user ?? null);
+      setSession(session);
+      if (session?.user) {
+        updatePushToken();
+      }
     });
 
-    return () => subscription.unsubscribe();
+    // クリーンアップ関数
+    return () => {
+      subscription.unsubscribe();
+    };
   }, []);
 
-  return { user };
+  return { session };
 }
